resolved cyclic import between device_utils and dut_ports modules
