+++ tests/common/helpers/dut_ports.py
@@ -1,6 +1,3 @@
-from tests.common.platform.device_utils import fanout_switch_port_lookup
-
-
 def encode_dut_port_name(dutname, portname):
     return dutname + '|' + portname
 
@@ -17,28 +14,3 @@ def decode_dut_port_name(dut_portname):
         dutname = None
         portname = None
     return dutname, portname
-
-
-def list_dut_fanout_connections(dut, fanouthosts):
-    """
-    Lists connected dut-fanout ports
-
-    Args:
-        dut: DUT host object
-        fanouthosts: List of fanout switch instances.
-
-    Returns:
-        A list of tuple with DUT's port, fanout port
-        and fanout
-    """
-    candidates = []
-
-    status = dut.show_interface(command='status')['ansible_facts']['int_status']
-
-    for dut_port in status.keys():
-        fanout, fanout_port = fanout_switch_port_lookup(fanouthosts, dut.hostname, dut_port)
-
-        if fanout and fanout_port and status[dut_port]['admin_state'] != 'down':
-            candidates.append((dut_port, fanout, fanout_port))
-
-    return candidates
