+++ tests/common/platform/device_utils.py
@@ -47,3 +47,28 @@ def get_dut_psu_line_pattern(dut):
         """
         psu_line_pattern = re.compile(r"PSU\s+(\d+).*?(OK|NOT OK|NOT PRESENT)\s+(green|amber|red|off)")
     return psu_line_pattern
+
+
+def list_dut_fanout_connections(dut, fanouthosts):
+    """
+    Lists connected dut-fanout ports
+
+    Args:
+        dut: DUT host object
+        fanouthosts: List of fanout switch instances.
+
+    Returns:
+        A list of tuple with DUT's port, fanout port
+        and fanout
+    """
+    candidates = []
+
+    status = dut.show_interface(command='status')['ansible_facts']['int_status']
+
+    for dut_port in status.keys():
+        fanout, fanout_port = fanout_switch_port_lookup(fanouthosts, dut.hostname, dut_port)
+
+        if fanout and fanout_port and status[dut_port]['admin_state'] != 'down':
+            candidates.append((dut_port, fanout, fanout_port))
+
+    return candidates
